pyproject.toml
[project]
name = "guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.0.20",
    "langchain-groq>=1.1.1",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "nemoguardrails[all]>=0.19.0",
    "python-dotenv>=1.2.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.8.0",
    "openai>=2.14.0",
    "pypdf2>=3.0.1",
    "rank-bm25>=0.2.2",
]
